// src/app/models/DetailFurnitureResponse.ts
export interface DetailFurnitureResponse {
    borrado: number
    colorFondo: string
    colorTexto: string
    conEtiqueta: number
    fbespacios: number
    fixelId: number
    fixeldaux: string
    id: number
    nivel: number
    plnid: number
    productos: string
    tipo_accesorio: string
    tramo: number
    esGranel : number;
  }