// src/app/models/FurnituresOpsHist.ts
export interface FurnituresOpsHist {
  tienda: number
  evento: number
  plnId: number
  depto: number
  numueble: number
  pacara: string
  mueble: string
  categorias: string
  fechainicio: string
  fechafin: string
  tiempo: string
  tramos: number
  fbalarmatiempo: number
  gerencia: string
  nombreTienda : string;
  departamento : string;
  layout: string
}